requirements/requirements.txt
# Core Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Essential Azure SDKs (only what you actually use)
azure-identity==1.25.0
azure-mgmt-resource==24.0.0
azure-mgmt-containerservice==39.1.0
azure-mgmt-monitor==7.0.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-compute==30.0.0
azure-mgmt-network==23.0.0
azure-mgmt-loganalytics==13.0.0b6
azure-mgmt-applicationinsights==4.0.0
azure-mgmt-consumption==10.0.0
azure-monitor-query==2.0.0
azure-core==1.35.1
azure-mgmt-core>=1.5.0
azure-common==1.1.28

# Data Processing (essential only)
pandas==2.0.3
numpy==1.24.3

# Machine Learning (for optimization algorithms)
scikit-learn==1.3.0
joblib==1.3.2

# HTTP & Utilities
requests==2.31.0
urllib3==2.0.4
python-dotenv==1.0.0
pyyaml==6.0.1

# Security
cryptography==41.0.3
PyJWT==2.8.0

# Database
sqlalchemy==2.0.19

# Basic utilities
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.0.3

# Template engine (for web UI)
jinja2==3.1.2

# Production WSGI Server (for Railway deployment)
gunicorn==21.2.0

# FastAPI (v3 — runs alongside Flask during migration)
fastapi>=0.115.0
uvicorn[standard]>=0.25.0
python-jose[cryptography]>=3.3.0
sse-starlette>=1.6.5
python-multipart>=0.0.6